Close #1463. -check reviewed.
