+++ src/lua/terrame.lua
@@ -1154,138 +1154,6 @@ function _Gtme.execProject(project, packageName)
 	return success, _
 end
 
-local function upperFirst(str)
-	return str:gsub("^%l", string.upper)
-end
-
-local function checkFile(file, prefixMsg)
-	local luacheck = require("luacheck.init")
-	local files = {file}
-	local options = {std = "min", cache = true, global = false}				
-	local issues = luacheck.check_files(files, options)
-	
-	if (issues.errors == 0) and (issues.fatals == 0) then	
-		issues = issues[1]
-		for _, issue in ipairs(issues) do
-			print(prefixMsg..": "..upperFirst(luacheck.get_message(issue))..". In file "..file..", line "..issue.line..".")
-		end		
-		
-		return #issues
-	end
-	
-	return 0
-end
-
-local function getLuaFiles(dirPath)
-	local files = {}
-	if Directory(dirPath):exists() then
-		_Gtme.forEachFile(dirPath, function(fname)
-			local fullPath = dirPath.."/"..fname
-			if Directory(fullPath):exists() then
-				for _, v in ipairs(getLuaFiles(fullPath)) do
-					table.insert(files, v)
-				end			
-			else 
-				local file = File(fullPath)
-				if file:extension() == "lua" then
-					table.insert(files, fullPath)
-				end
-			end
-		end)
-	end
-	
-	return files
-end
-
-local function getRelativePath(full, absoluteLength)
-	return string.sub(full, absoluteLength + 2)
-end
-
-local function checkPackage(package, packagePath)
-	_Gtme.printNote("Running code analyzer for package '"..package.."'")
-	local clock0 = os.clock()
-	
-	local testsPath = packagePath.."/tests"
-	local luaPath = packagePath.."/lua"
-	local testFiles = getLuaFiles(testsPath)
-	local luaFiles = getLuaFiles(luaPath)	
-	
-	local srcPath
-	local srcFiles = {}
-	local srcPathLenght
-	if package == "base" then
-		srcPath = sessionInfo().path.."/lua"
-		srcFiles = getLuaFiles(srcPath)
-		srcPathLenght = string.len(srcPath) 
-	end
-	
-	local luacheck = require("luacheck.init")	
-	local numIssues = 0
-	local pkgPathLenght = string.len(packagePath)
-	local options = {std = "min", cache = true, global = false}	
-	
-	_Gtme.printNote("Analysing source code")
-	for _, file in ipairs(luaFiles) do
-		local files = {file}
-		local issues = luacheck.check_files(files, options)[1]
-		for _, issue in ipairs(issues) do
-			_Gtme.printError("Warning: "..upperFirst(luacheck.get_message(issue))..". In file "..getRelativePath(file, pkgPathLenght)..", line "..issue.line..".")
-		end	
-		numIssues = numIssues + #issues
-	end
-
-	if #srcFiles > 0 then
-		for _, file in ipairs(srcFiles) do
-			local files = {file}
-			local issues = luacheck.check_files(files, options)[1]
-			for _, issue in ipairs(issues) do
-				_Gtme.printError("Warning: "..upperFirst(luacheck.get_message(issue))..". In file "..getRelativePath(file, srcPathLenght)..", line "..issue.line..".")
-			end	
-			numIssues = numIssues + #issues
-		end	
-	end
-	
-	_Gtme.printNote("Analysing tests")
-	for _, file in ipairs(testFiles) do
-		local files = {file}
-		local issues = luacheck.check_files(files, options)[1]
-		for _, issue in ipairs(issues) do
-			_Gtme.printError("Warning: "..upperFirst(luacheck.get_message(issue))..". In file "..getRelativePath(file, pkgPathLenght)..", line "..issue.line..".")
-		end	
-		numIssues = numIssues + #issues
-	end
-	
-	_Gtme.printNote("\nCode analyzer report for package '"..package.."':")
-	
-	local clock1 = os.clock()
-	local dt = clock1 - clock0
-	_Gtme.printNote("Analysis executed in "..round(dt, 2).." seconds.")
-	
-	local totalFiles = #testFiles + #luaFiles + #srcFiles
-	
-	if totalFiles > 0 then
-		if totalFiles == 1 then
-			_Gtme.printNote("One file was analysed.")
-		else
-			_Gtme.printNote(totalFiles.." files were analysed.")
-		end
-		
-		if numIssues > 0 then
-			if numIssues == 1 then
-				_Gtme.printNote("One issue was found.")
-			else
-				_Gtme.printNote(numIssues.." issues were found.")
-			end
-		else
-			_Gtme.printNote("Success, no issue found.")
-		end
-	else
-		_Gtme.printNote("No file found.")
-	end
-	
-	return numIssues
-end
-
 function _Gtme.execute(arguments) -- 'arguments' is a vector of strings
 	info_ = { -- this variable is used by Utils:sessionInfo()
 		mode = "normal",
@@ -1643,6 +1511,8 @@ function _Gtme.execute(arguments) -- 'arguments' is a vector of strings
 					os.exit(0)
 				end
 			elseif arg == "-check" then
+				dofile(sessionInfo().path.."lua"..s.."check.lua")
+
 				if info_.package == nil then
 					info_.package = "base"
 				end
@@ -1653,7 +1523,7 @@ function _Gtme.execute(arguments) -- 'arguments' is a vector of strings
 					os.exit(1)
 				end)
 
-				local numIssues = checkPackage(package, pkgPath)
+				local numIssues = _Gtme.checkPackage(package, pkgPath)
 			
 				os.exit(numIssues)				
 			else
@@ -1697,10 +1567,12 @@ function _Gtme.execute(arguments) -- 'arguments' is a vector of strings
 
 			info_.currentFile = arg
 
+			dofile(sessionInfo().path.."lua"..s.."check.lua")
+
 			if info_.mode == "strict" then
-				checkFile(arg, "Warning")
+				_Gtme.checkFile(arg, "Warning")
 			elseif info_.mode == "debug" then
-				local numIssues = checkFile(arg, "Error")
+				local numIssues = _Gtme.checkFile(arg, "Error")
 				if numIssues > 0 then
 					os.exit(numIssues)
 				end
